fix compiler warnings in hurd/console-client
console-client: Fix several compiler warnings.

* console-client/bdf.c: Use size_t instead of int.
* console-client/driver.c: Remove unused variable errstring.
* console-client/pc-kbd.c: Cast sc to scancode_x1 before comparing with enum
values.
* console-client/vga-dynacolor.h: Use an explicit if in reference counting.
* console-client/vga-dynafont.c: Use usigned char for bitmaps.
* console-client/vga-support.c: Use unsigned char instead of char.
* console-client/vga-support.h: Likewise.
* console-client/vga.c: Use conchar_attr_equal instead of casting structures to
integers.
* hurd/console.h: Add conchar_attr_equal to compare conchar_attr_t structures.

+++ console-client/bdf.c
@@ -128,7 +128,7 @@ parse_hexbyte (char *line, unsigned char *byte)
    COMMENT lines, and removes whitespace at the beginning and end of a
    line.  */
 static int
-next_line (char **line, int *size, FILE *file, int *count)
+next_line (char **line, size_t *size, FILE *file, int *count)
 {
   int len;
 
@@ -184,7 +184,7 @@ bdf_read (FILE *filep, bdf_font_t *font, int *linecount)
 {
   bdf_error_t err = 0;
   char *line = 0;
-  int line_size = 0;
+  size_t line_size = 0;
   int len;
   int done = 0;
   bdf_font_t bdf;
